Actualizacion de la API

+++ src/main/java/com/example/proyectoServidor/service/DeveloperServiceImpl.java
@@ -2,7 +2,6 @@ package com.example.proyectoServidor.service;
 
 import com.example.proyectoServidor.model.Developer;
 import com.example.proyectoServidor.repository.DeveloperRepository;
-import com.example.proyectoServidor.service.DeveloperService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -31,11 +30,22 @@ public class DeveloperServiceImpl implements DeveloperService {
 
     @Override
     public Developer saveDeveloper(Developer developer) {
+        if (developerRepository.existsByEmail(developer.getEmail())) {
+            throw new RuntimeException("Email already exists");
+        }
         return developerRepository.save(developer);
     }
 
     @Override
     public void deleteDeveloper(Integer developerId) {
+        if (!developerRepository.existsById(developerId)) {
+            throw new RuntimeException("Developer not found");
+        }
         developerRepository.deleteById(developerId);
+    }
+
+    @Override
+    public boolean existsByEmail(String email) {
+        return developerRepository.existsByEmail(email);
     }
 }
